pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "implicitgeo"
version = "0.1.0"
dependencies = ["numpy", "scikit-image", "numpy-stl"]
